requirements.txt
# Feed RSS parser
feedparser==6.0.10

# Database as JSON
pickleDB==0.9.2

# Scheduler (cron-like) for Python
APScheduler==3.10.3

# Logging
loguru==0.7.0

# YAML parser
PyYAML==6.0.1

# Notifications
apprise==1.4.5

# JSON transformer
json_dict_transformer==1.1.1

# YouTube API
pytube==15.0.0

# Todoist API
todoist-api-python==2.1.3
